Modified snapshot watermark and log header to identify as a Rescale Experiment build.
If people are going to be swapping rescale executables in and out of trunk installations, there's a small chance that problems caused by rescaling changes might inadvertently be reported as bugs in a trunk build. If the watermark and log header clearly identify it as an experimental build, the chances of this happening are reduced.